--- a/news/3 Code Health/12027.md
+++ b/news/3 Code Health/12027.md
@@ -1,0 +1,1 @@
+Faster node unit tests on Azure pipeline.