--- a/src/test/datascience/ipywidgets/cdnWidgetScriptSourceProvider.unit.test.ts
+++ b/src/test/datascience/ipywidgets/cdnWidgetScriptSourceProvider.unit.test.ts
@@ -40,11 +40,13 @@ suite('DataScience - ipywidget - CDN', () => {
     let fileSystem: IFileSystem;
     let webviewUriConverter: ILocalResourceUriConverter;
     let tempFileCount = 0;
-    setup(function () {
+    suiteSetup(function () {
         // Nock seems to fail randomly on CI builds. See bug
         // https://github.com/microsoft/vscode-python/issues/11442
         // tslint:disable-next-line: no-invalid-this
-        this.skip();
+        return this.skip();
+    });
+    setup(() => {
         notebook = mock(JupyterNotebookBase);
         configService = mock(ConfigurationService);
         httpClient = mock(HttpClient);