For #11999
* When running unit tests we don't need to compile the webviews (no need to compile react, etc).
* Total time to compile code down from 6 minutes to 1 minute (i.e. 5 minutes faster CI to know whether tests passed/failed)

Note the drop in compile times from ~7minutes to ~50s. (**basically 6 minutes faster**)

| Before        | After           | 
| ------------- |:-------------:| 
| ![Screen Shot 2020-05-27 at 16 31 19](https://user-images.githubusercontent.com/1948812/83082224-a760b600-a037-11ea-807f-01666fc2f157.png) | ![Screen Shot 2020-05-27 at 16 31 34](https://user-images.githubusercontent.com/1948812/83082231-ad569700-a037-11ea-9bd8-38b8110ac70b.png) |